Optimize imported data from int32 to uint8 Before there was 4096 int32's being passed into the shader, which is 16384 bytes Now there is 4096 uint8's which is 4096 bytes While this does limit the amount of ID's to just 256 (including 0), because the shader renders just 1 chunk, and it's planned to have many more chunks, it's better to reduce memory that is being passed to GPU, I think.

+++ source/world.py
@@ -14,7 +14,7 @@ class Chunk:
 
     def __init__(self, position: tuple[int, int, int]):
         self.position: tuple[int, int, int] = position
-        self.voxels: np.ndarray = np.zeros([CHUNK_SIZE, CHUNK_SIZE, CHUNK_SIZE], dtype=np.int32)
+        self.voxels: np.ndarray = np.zeros([CHUNK_SIZE, CHUNK_SIZE, CHUNK_SIZE], dtype=np.uint8)
 
 
 def generate_flat(level: int) -> Chunk:
